module rename + import change in test.py
